refactor(#3272): extract closure-exports from index.ts (refs #3182)
__call_fn_<N>/__call_fn_method_<N> host-dispatch + closure classification
exports moved verbatim; index.ts imports back the 10 driver entry points.
Byte-identity IDENTICAL (39/39), tsc 0.

✓
